Improve install command (#94)
* feat(install): add network policies in ClusterRole

Signed-off-by: Elis Lulja <elulja@cisco.com>

* feat(install): ask for `watchAllNetworkPolicies`

Signed-off-by: Elis Lulja <elulja@cisco.com>

* feat(install): pull latest version from github

Signed-off-by: Elis Lulja <elulja@cisco.com>

* refactor(install): improve cleaning up process

Signed-off-by: Elis Lulja <elulja@cisco.com>

* refactor(install): create `askYesNo` func

This is a utility function for avoiding repetition for questions
that need a yes or no.

Signed-off-by: Elis Lulja <elulja@cisco.com>

* feat(install): create `installer` struct

Signed-off-by: Elis Lulja <elulja@cisco.com>

* refactor(install): use `install` from `installer`

Signed-off-by: Elis Lulja <elulja@cisco.com>

* refactor(install): generate names

This commit generates name -- e.g. the `ConfigMap` or `Secret` name --
starting from the name that is given to the deployment.

Signed-off-by: Elis Lulja <elulja@cisco.com>

* refactor(yaml): update resources names

This commit changes the name of the resources -- e.g. the
`ServiceAccount` etc -- to reflect the changes that were done in the
previous commit.

Signed-off-by: Elis Lulja <elulja@cisco.com>

* refactor(install): better clean up

The `cleanUp` function has been rewritten to avoid repetition.

Signed-off-by: Elis Lulja <elulja@cisco.com>

* refactor(install): use same flag names as `run`

Signed-off-by: Elis Lulja <elulja@cisco.com>

* refactor(install): create `askForPassword`

This commit creates a `askForPassword` function because it is used both
by the interactive installer and the flags-based one. So that no trace
of the password is left on the user's device.

Signed-off-by: Elis Lulja <elulja@cisco.com>

* feat(install): add `--image` flag

Signed-off-by: Elis Lulja <elulja@cisco.com>

---------

Signed-off-by: Elis Lulja <elulja@cisco.com>